econ_scope_webapp/econ_scope/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
fonttools==4.47.2
frozenlist==1.4.1
g4f==0.2.0.7
idna==3.6
kiwisolver==1.4.5
lsd==0.3.3
matplotlib==3.8.2
multidict==6.0.4
mysqlclient==2.2.1
numpy==1.26.3
ocrd-fork-pylsd==0.0.8
opencv-python-headless==4.9.0.80
packaging==23.2
pika==1.3.2
pillow==10.2.0
pylsd==0.0.2
pyparsing==3.1.1
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2023.4
requests==2.31.0
scipy==1.12.0
six==1.16.0
sqlparse==0.4.4
urllib3==2.1.0
yarl==1.9.4
